Test api key on first request
