+++ test/integration-tests/trtest2.js
@@ -3,7 +3,7 @@
 // (Second block)
 //
 
-var assert = require('./trassert');
+var assert = require('../trassert');
 
 var api = require('../../api');
